requirements.txt
Flask==2.1.2
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
mysql-connector-python==8.0.28
